MINOR: rsyslog: Fix Oracle audit logs send with spaces in syslog tag
Seen at least with an Oracle running on AIX, audit and unified audit logs
ares sent respectively with "Oracle Audit" and "Oracle Unified Audit" syslog
tags. Those contain spaces, which is not RFC compliant, causing mess with
syslog parser for standard fields (host, tag).

This is fix is for the case when AIX option -n is used to supress the "message
forwarded from <host>: " prefix, which was already fixed.