MEDIUM: wui: Use vue router hash mode instead of history mode
History mode is not compatible with a relative base. This is an ugly
hack that breaks conventions of what an URL is. It requires server-side
tricks to serve the app index.html on any path, which breaks 404, ...,
also I could not find a proper way to setup that base url at runtime
so it can be generic without rebuilding the app.

createWebHistory with relative path #1245
https://github.com/vuejs/router/issues/1245