MINOR: keepalived: Add configuration for web frontal VIP
This is bascically the same as existing keepalived for kibana VIP.
I prefer not to rename the kibana files for now to avoid a breaking
change on update. The more generic term web is used instead of kibana
because we have now another web interface for raw logs archives which
can be used is some cases.